Appendix/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
snowflake-connector-python==3.0.0
python-dotenv==1.0.0
